etu_django_frame/django/3.2.13/requirements/requirement.txt
Django==3.2.13
celery==5.1.2
django-cas-ng==4.1.0
django-celery-beat==2.2.1
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-redis-cache==3.0.1
django-redis-sessions==0.6.2
djangorestframework==3.12.4
django-simpleui==2021.8.17
drf-yasg==1.20.0
Markdown==3.2.1
ipython==7.16.3
kombu==5.1.0
mysqlclient==2.0.3
redis==3.5.3
requests==2.25.1
supervisor==4.2.2
uWSGI==2.0.20
vine==5.0.0